Moved code from AccelerationGlobal to ExtensionGlobal
